Extract ImportTokenModal from Swap component
